+++ sdk/base/nostd/traits.hpp
@@ -161,9 +161,9 @@ template <> struct make_unsigned<uint128> { using type = uint128; };
 
 enum class type_category_t {
     unknown,
-    integral,        // int, char, bool (std::is_integral) + int64, uint64 (custom::is_integral)
-    floating_point,  // float, double (std::is_floating_point)
-    cstring          // std::string, char*
+    integral,  // int, char, bool (std::is_integral) + int64, uint64 (custom::is_integral)
+    real,      // float, double (std::is_floating_point)
+    literal,   // std::string, char*
 };
 
 template <typename T>
@@ -171,9 +171,9 @@ struct get_type_category {
     using decay_t = typename std::decay<T>::type;
 
     static const type_category_t value =  //
-        custom::is_integral<decay_t>::value      ? type_category_t::integral
-        : std::is_floating_point<decay_t>::value ? type_category_t::floating_point
-        : (std::is_same<decay_t, const char*>::value || std::is_same<decay_t, char*>::value || std::is_same<decay_t, std::string>::value) ? type_category_t::cstring
+        custom::is_integral<decay_t>::value                                                                                               ? type_category_t::integral
+        : std::is_floating_point<decay_t>::value                                                                                          ? type_category_t::real
+        : (std::is_same<decay_t, const char*>::value || std::is_same<decay_t, char*>::value || std::is_same<decay_t, std::string>::value) ? type_category_t::literal
                                                                                                                                           : type_category_t::unknown;
 };