+++ sdk/base/nostd/types.hpp
@@ -15,7 +15,19 @@
 
 namespace hotplace {
 
-//
+enum class range_type_t : int8 {
+    minvalue = -1,    // -inf
+    ninf = minvalue,  // negative inf
+    value = 0,        // value
+    maxvalue = 1,     // +inf
+    inf = maxvalue,   // positive inf
+};
+enum class range_flag_t : uint8 {
+    excluded = 0,       // open
+    open = excluded,    //
+    included = 1,       // closed
+    closed = included,  //
+};
 
 }  // namespace hotplace