+++ sdk/base/unittest/testcase.hpp
@@ -7,7 +7,8 @@
  * Revision History
  * Date         Name                Description
  * 2017.07.13   Soo Han, Kim        prototype (codename.grape Revision 288)
- * 2023.08.15   Soo Han, Kim        elapsed time (codename.hotplace Revision 9)
+ * 2023.08.15   Soo Han, Kim        reboot (codename.hotplace Revision 9)
+ *
  */
 
 #ifndef __HOTPLACE_SDK_BASE_UNITEST_TESTCASE__