+++ test/testcase/io/parser/testcase_parser.cpp
@@ -284,8 +284,10 @@ void test_lalr() {
 
             // 3. Primary Elements (SIZE, FROM, PATTERN, Range, Parenthesized)
             .add_production("PrimaryElement", {"ValueElement"})
-            .add_production("PrimaryElement", {"ValueElement", "..", "ValueElement"})
-            .add_production("PrimaryElement", {"ValueElement", "..", "<", "ValueElement"})
+            .add_production("PrimaryElement", {"ValueElement", "..", "ValueElement"})            // [from, to]
+            .add_production("PrimaryElement", {"ValueElement", "..", "<", "ValueElement"})       // [from, to)
+            .add_production("PrimaryElement", {"ValueElement", "<", "..", "ValueElement"})       // (from, to]
+            .add_production("PrimaryElement", {"ValueElement", "<", "..", "<", "ValueElement"})  // (from, to)
             .add_production("PrimaryElement", {"SIZE", "Constraint"})
             .add_production("PrimaryElement", {"FROM", "Constraint"})
             .add_production("PrimaryElement", {"PATTERN", symqs})
@@ -313,6 +315,7 @@ void test_lalr() {
             .add_terminal("]")
             .add_terminal("(")
             .add_terminal(")")
+            .add_terminal("<")
             .add_terminal("..")
             .add_terminal("|")
             .add_terminal("INTERSECTION")