+++ CHANGELOG.md
@@ -35,6 +35,10 @@
     * CBOR
 
 * details
+  * Revision 1082
+    * asn1_publisher - Constraints
+      * exclusive boundary (1..<256, 1<..256, 1<..<256)
+      * string range ("a".."z")
   * Revision 1081
     * asn1_publisher - Constraints
   * Revision 1079